requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.5
django-redis-cache==1.7.1
idna==2.6
kombu==4.1.0
oauthlib==2.0.3
psycopg2==2.7.3.1
PyJWT==1.5.3
python-dateutil==2.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
slackclient==1.0.9
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
websocket-client==0.44.0
